Move probability functions into seperate file
